pyproject.toml
[tool.poetry]
name = "langgraph-cookbook"
version = "0.0.1"
description = ""
authors = ["ibrahimsaidi <ibrahim.aka.ajax@gmail.com>"]
readme = "README.md"
packages =  [
    { include = "apps", from = "." },
    { include = "tools", from = "." },
    { include = "utils", from = "." }
]
[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.1.5"
langchain-core = "^0.2.10"
langchain = "^0.2.6"
langchain-experimental = "^0.0.62"
langchain-openai = "^0.1.13"
resend = "^2.1.0"
exa-py = "^1.0.12"
supabase = "^2.5.1"
python-dotenv = "^1.0.1"
modal = "^0.63.31"
pandas = "^2.2.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
